refactor: extract base config from test file
